+++ src/Workspaces/Core/Portable/LanguageServices/SyntaxFactsService/AbstractSyntaxFactsService.cs
@@ -7,11 +7,8 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
 using System.Linq;
-using System.Text;
 using System.Threading;
-using Microsoft.CodeAnalysis.FindSymbols;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Shared.Utilities;
@@ -20,95 +17,6 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.LanguageServices
 {
-    internal abstract class AbstractDeclaredSymbolInfoFactoryService : IDeclaredSymbolInfoFactoryService
-    {
-        private const string GenericTypeNameManglingString = "`";
-        private static readonly string[] s_aritySuffixesOneToNine = { "`1", "`2", "`3", "`4", "`5", "`6", "`7", "`8", "`9" };
-
-        private readonly static ObjectPool<List<Dictionary<string, string>>> s_aliasMapListPool
-            = SharedPools.Default<List<Dictionary<string, string>>>();
-
-        // Note: these names are stored case insensitively.  That way the alias mapping works 
-        // properly for VB.  It will mean that our inheritance maps may store more links in them
-        // for C#.  However, that's ok.  It will be rare in practice, and all it means is that
-        // we'll end up examining slightly more types (likely 0) when doing operations like 
-        // Find all references.
-        private readonly static ObjectPool<Dictionary<string, string>> s_aliasMapPool
-            = SharedPools.StringIgnoreCaseDictionary<string>();
-
-        protected static List<Dictionary<string, string>> AllocateAliasMapList()
-            => s_aliasMapListPool.Allocate();
-
-        protected static void FreeAliasMapList(List<Dictionary<string, string>> list)
-        {
-            if (list != null)
-            {
-                foreach (var aliasMap in list)
-                {
-                    FreeAliasMap(aliasMap);
-                }
-
-                s_aliasMapListPool.ClearAndFree(list);
-            }
-        }
-
-        protected static void FreeAliasMap(Dictionary<string, string> aliasMap)
-        {
-            if (aliasMap != null)
-            {
-                s_aliasMapPool.ClearAndFree(aliasMap);
-            }
-        }
-
-        protected static Dictionary<string, string> AllocateAliasMap()
-            => s_aliasMapPool.Allocate();
-
-        protected static void AppendTokens(SyntaxNode node, StringBuilder builder)
-        {
-            foreach (var child in node.ChildNodesAndTokens())
-            {
-                if (child.IsToken)
-                {
-                    builder.Append(child.AsToken().Text);
-                }
-                else
-                {
-                    AppendTokens(child.AsNode(), builder);
-                }
-            }
-        }
-
-        protected static void Intern(StringTable stringTable, ArrayBuilder<string> builder)
-        {
-            for (int i = 0, n = builder.Count; i < n; i++)
-            {
-                builder[i] = stringTable.Add(builder[i]);
-            }
-        }
-
-        public static string GetMetadataAritySuffix(int arity)
-        {
-            Debug.Assert(arity > 0);
-            return (arity <= s_aritySuffixesOneToNine.Length)
-                ? s_aritySuffixesOneToNine[arity - 1]
-                : string.Concat(GenericTypeNameManglingString, arity.ToString(CultureInfo.InvariantCulture));
-        }
-
-        public abstract bool TryGetDeclaredSymbolInfo(StringTable stringTable, SyntaxNode node, string rootNamespace, out DeclaredSymbolInfo declaredSymbolInfo);
-
-        /// <summary>
-        /// Get the name of the target type of specified extension method declaration. 
-        /// The node provided must be an extension method declaration,  i.e. calling `TryGetDeclaredSymbolInfo()` 
-        /// on `node` should return a `DeclaredSymbolInfo` of kind `ExtensionMethod`. 
-        /// If the return value is null, then it means this is a "complex" method (as described at <see cref="SyntaxTreeIndex.ExtensionMethodInfo"/>).
-        /// </summary>
-        public abstract string GetTargetTypeName(SyntaxNode node);
-
-        public abstract bool TryGetAliasesFromUsingDirective(SyntaxNode node, out ImmutableArray<(string aliasName, string name)> aliases);
-
-        public abstract string GetRootNamespace(CompilationOptions compilationOptions);
-    }
-
     internal abstract class AbstractSyntaxFactsService
     {
         private readonly static ObjectPool<Stack<(SyntaxNodeOrToken nodeOrToken, bool leading, bool trailing)>> s_stackPool
